// src/typings/enums.ts
export enum TimesUnits {
    Year = "year",
    Month = "month",
    Week = "week",
    Day = "day",
    Hour = "hour",
    Minute = "minute",
}

export enum TimesInMilliseconds {
    Year = 31557600000,
    Month = 2629800000,
    Week = 604800000,
    Day = 86400000,
    Hour = 3600000,
    Minute = 60000,
}